handle selected icon-only tabs
"the calligra situation" - we make them
look a bit like a listview for uninverted
headers.
For inverted headers, we break the inversion
(the reason is also that calligra hardcodes the
text into the icon -this is btw. really silly,
they could have just bound a listview to a stack-
with the "default" color, ie. unreadable on inversion)